one more commit

+++ src/test/java/Operations.java
@@ -1,5 +1,4 @@
 import ru.yandex.qatools.allure.annotations.Parameter;
-import ru.yandex.qatools.allure.annotations.Step;
 
 /**
  * Created by liza on 26.09.15.
@@ -7,13 +6,13 @@ import ru.yandex.qatools.allure.annotations.Step;
 public abstract class Operations {
 
     @Parameter("First operand")
-    int operand1;
+    long operand1;
     @Parameter("Second operand")
-    int operand2;
+    long operand2;
     @Parameter("Expected result")
-    double result;
+    Double result;
 
-    public Operations(int operand1, int operand2, double result) {
+    public Operations(long operand1, long operand2, double result) {
         this.operand1 = operand1;
         this.operand2 = operand2;
         this.result = result;
